Revert "Issue #2977738 by chr.fritsch, mtodor: Install JSON-LD support and update to AMP 2.x release"
This reverts commit 0e90383ad93d21e1bcd597545f345d80be286af9.
