+++ composer.json
@@ -66,6 +66,9 @@
                 "create a NEW paragraph 'in place' between existing ones": "https://www.drupal.org/files/issues/2899917_6.patch",
                 "Schema errors": "https://www.drupal.org/files/issues/build_failing_schema-2881459-7.patch"
             },
+            "drupal/amp": {
+                "Missing schema": "https://www.drupal.org/files/issues/missing_schema_for-2878769-3.patch"
+            },
             "drupal/entity_browser": {
                 "Provide inline entity form FieldWidgetDisplay": "https://www.drupal.org/files/issues/2858438_6.patch"
             }
@@ -76,7 +79,7 @@
         "burdamagazinorg/infinite_theme": "^1.0",
         "composer/installers": "^1.0",
         "cweagans/composer-patches": "~1.0",
-        "drupal/amp": "^1.0||^2.0",
+        "drupal/amp": "1.1",
         "drupal/amptheme": "~1.0",
         "oomphinc/composer-installers-extender": "^1.1",
         "drupal/access_unpublished": "^1.0",
@@ -124,7 +127,6 @@
         "drupal/redirect": "^1.0",
         "drupal/riddle_marketplace": "^2.0",
         "drupal/scheduler": "^1.0",
-        "drupal/schema_metatag": "^1.0",
         "drupal/simple_sitemap": "^2.0",
         "drupal/shariff": "^1.0",
         "drupal/slick": "^1.0",
