+++ drupal-org.make
@@ -7,7 +7,8 @@ projects[admin_toolbar][version] = 1
 projects[adsense][type] = module
 projects[adsense][version] = 1
 projects[amp][type] = module
-projects[amp][version] = 2
+projects[amp][version] = 1.1
+projects[amp][patch][] = https://www.drupal.org/files/issues/missing_schema_for-2878769-3.patch
 projects[amptheme][type] = theme
 projects[amptheme][version] = 1
 projects[better_normalizers][type] = module
@@ -98,8 +99,6 @@ projects[redirect][type] = module
 projects[redirect][version] = 1
 projects[scheduler][type] = module
 projects[scheduler][version] = 1
-projects[schema_metatag][type] = module
-projects[schema_metatag][version] = 1
 projects[shariff][type] = module
 projects[shariff][version] = 1
 projects[simple_gmap][type] = module
