--- modules/thunder_article/config/update/thunder_article__thunder_update_8121.yml
@@ -1,14 +1,0 @@
-metatag.metatag_defaults.node__article:
-  expected_config: { }
-  update_actions:
-    add:
-      tags:
-        schema_article_about: '[node:field_channel]'
-        schema_article_date_modified: '[node:changed:html_datetime]'
-        schema_article_date_published: '[node:created:html_datetime]'
-        schema_article_description: '[node:field_teaser_text]'
-        schema_article_headline: '[node:field_seo_title]'
-        schema_article_image: 'a:2:{s:5:"@type";s:11:"ImageObject";s:3:"url";s:53:"[node:field_teaser_media:entity:field_image:facebook]";}'
-        schema_article_name: '[node:title]'
-        schema_article_publisher: 'a:3:{s:5:"@type";s:12:"Organization";s:4:"name";s:11:"[site:name]";s:3:"url";s:10:"[site:url]";}'
-        schema_article_type: NewsArticle
