+++ tests/src/FunctionalJavascript/ThunderMetaTagTrait.php
@@ -2,8 +2,6 @@
 
 namespace Drupal\Tests\thunder\FunctionalJavascript;
 
-use Drupal\Component\Serialization\Json;
-
 /**
  * Trait for manipulation of meta tag configuration and meta tags on page.
  *
@@ -59,22 +57,17 @@ trait ThunderMetaTagTrait {
    *   Meta tag value.
    */
   protected function checkMetaTag($name, $value) {
-    if (strpos($name, 'schema_article') === 0) {
-      $jsonLd = Json::decode($this->getSession()->getPage()->find('xpath', '//head/script[(@type="application/ld+json")]')->getHtml());
-      $key = substr($name, strlen('schema_article_'));
-      $this->assertSame($jsonLd['@graph'][0][$key], $value);
-    }
-    else {
-      $htmlValue = htmlentities($value);
-      $checkXPath = "@content='{$htmlValue}'";
-      if (strpos($value, 'LIKE:') === 0) {
-        $valueToCheck = substr($htmlValue, strlen('LIKE:'));
+    $htmlValue = htmlentities($value);
 
-        $checkXPath = "contains(@content, '{$valueToCheck}')";
-      }
-      $this->assertSession()
-        ->elementExists('xpath', "//head/meta[(@name='{$name}' or @property='{$name}') and {$checkXPath}]");
+    $checkXPath = "@content='{$htmlValue}'";
+    if (strpos($value, 'LIKE:') === 0) {
+      $valueToCheck = substr($htmlValue, strlen('LIKE:'));
+
+      $checkXPath = "contains(@content, '{$valueToCheck}')";
     }
+
+    $this->assertSession()
+      ->elementExists('xpath', "//head/meta[(@name='{$name}' or @property='{$name}') and {$checkXPath}]");
   }
 
   /**
