+++ thunder.info.yml
@@ -66,7 +66,6 @@ dependencies:
   - paragraph_split_text
   - pathauto
   - redirect
-  - schema_article
   - scheduler
   - simple_sitemap
   - slick_media
