+++ thunder.install
@@ -944,20 +944,3 @@ function thunder_update_8120() {
   // Output logged messages to related channel of update execution.
   return $thunderUpdater->logger()->output();
 }
-
-/**
- * Install JSON-LD support.
- */
-function thunder_update_8121() {
-  /** @var \Drupal\thunder_updater\Updater $thunderUpdater */
-  $thunderUpdater = \Drupal::service('thunder_updater');
-
-  if ($thunderUpdater->installModules(['thunder_article__thunder_update_8121' => 'schema_article'])) {
-    if (!$thunderUpdater->executeUpdates([['thunder_article', 'thunder_article__thunder_update_8121']])) {
-      $thunderUpdater->checklist()->markUpdatesFailed(['thunder_article__thunder_update_8121']);
-    }
-  }
-
-  // Output logged messages to related channel of update execution.
-  return $thunderUpdater->logger()->output();
-}
